+++ src/ipc_actor_define_common.h
@@ -5,6 +5,7 @@ struct ipc_actor;
 struct ipc_msg;
 
 #ifdef __cplusplus
+#define _IPC_STATIC_ASSERT(condition, message) static_assert((condition), message)
 extern "C" {
 #endif
 
@@ -33,7 +34,7 @@ void ipc_dispatch_actor_handlers(struct ipc_actor *self, const struct ipc_msg *m
         handler_fn(self, (const MsgType##_payload_t *) payload, raw_msg);           \
     }
 
-#define IPC_ACTOR_RESPONSE_HANDLE(actor_sym, RequestType, ReplyType, handler_fn)            \
+#define _IPC_ACTOR_RESPONSE_HANDLE_IMPL(actor_sym, RequestType, ReplyType, handler_fn)      \
     static_assert(sizeof(ReplyType##_payload_t) <= actor_sym##_max_payload_size,            \
                   #ReplyType " reply payload exceeds actor max payload size");              \
     static void handler_fn##_ipc_typed(struct ipc_actor *self, int result,                  \
@@ -59,4 +60,6 @@ void ipc_dispatch_actor_handlers(struct ipc_actor *self, const struct ipc_msg *m
 
 #ifdef __cplusplus
 }
+#else
+#define _IPC_STATIC_ASSERT(condition, message) _Static_assert((condition), message)
 #endif